renamed 'elements' to 'constituent terms'/'constituents'
There was an ambiguity with graphs, as they are defined as sets of triples.
The term 'element' in that context could refer to 'elements of the set'
(i.e. triples), or 'elements of its triples' (i.e. any kinds of terms).

Even worse, 'asserted triples' are the triples that are (set-)elements of
the graph, while 'embedded triples' are the triples that are constituents
of the graph. Calling both "elements" could therefore cause much confusion.