(yet another) semantics proposal
this one is actually very close to what @pfps proposed a while ago:
https://www.w3.org/mid/3ae4b60f-7de6-e552-1443-657110ab0d30@gmail.com

Note that in this proposal, unstar(G) is NOT equivalent to G
(except in some special cases).
This removes the need to have hidden IRIs,
and ensures (I think) that SPARQL's evaluation semantics is consistent
with simple entailment.