reorganized remarks about Semantics
Now that TEPs are part of the normative text,
comments about it being "possible but not specified in this document"
are moot.

The section about 'Selective ref transparency' has been moved up,
right after the one about 'Ref opacity', with a smoother transition
(former parts of 'Ref opacity' are now the intro of
'Selective ref transparency').